Adding perl version of versionfinder. From my tests this runs about 10x faster with the test cases I have used.
Does not support all of the old flags since many of them are legacy and
output changes are handled inside the code now.

Some of the missing features may be added in if requested.

Will be removing the old bash version as soon as I have done some
additional testing.